Bugfixes, control flow obfuscation, code cleanup
Fixed bug where predicted size of instruction would be wrong when updating the code reference.

Fixed another bug where the values wouldn't overflow in the linear substitution, leading to operands being too big in size. This is fixed by overflowing manually.

Added mnemonics and registers to the list.

Added control flow obfuscation - this adds an 0xEB byte before each instruction that starts with 0xFF byte. This makes a jmp FF which ends up executing original instruction, but will make disassembly harder.

Added a 'rsp relative' method that checks if any of the operands in a disassembled instruction use the rsp register. This is used to skip instructions for obfuscation if they are rsp relative (to avoid clobbering the stack).