Added control flow flattening
This takes a function's basic blocks and assigns them each their own unique index.

To execute a basic block, the index of the block is passed to a stub which will jump to the correct relative address.

This means that all basic blocks have to be redirected to jump to that stub before moving onto the next block.

The 'target' basic block is the block that will be ran if a conditional jump (or unconditional jump) is true.

The 'fallthrough' block is the block that will execute if an unconditional jump's statement is false OR if there is no control flow operation at the end of the block (goes to next instruction).

The 'fallthrough' and 'target' branches are redirected to the stub with their correct indexes.

The basic blocks of the functions are also shuffled to prevent recovery of control flow.