Added CPU flag emulation to mixed boolean arithmetic
This works by recreating the flag behavior that would have taken place with the original arithmetic operation.

When mutating with MBA, the flags will not be the same as they would have been.

For 'xor', 'or', 'and' the following instruction is inserted after the MBA operation:

test result, result

This correctly simulates the behavior for those instructions.

For the 'sub x, y' instruction, the following is inserted:

cmp x, y

This works because a 'cmp' instruction modifies the flags in the exact same manner as 'sub'.

For the 'add' instruction:

A 'test result, result' instruction is added for the parity flag, sign flag, and zero flag

The auxiliary carry flag, carry flag, and overflow flag have to be computed manually (and set/unset in the rflags) using a stub we insert at the end of the operation.