Improved CPU flags emulation in MBA
Now it will only run the flag emulation if there is an instruction that reads the CPU flags near to it or if it is the last flag writing instruction in the block.

This avoids having to have many emulations of the flags unnecessarily (which could hint to what the original instruction was).