Added virtual machine obfuscation
This takes an instruction and places it within a virtualized context.

To enter the virtual machine, we push flags and all general purpose registers onto the stack.

The stack layout for the general purpose registers is randomized for every entry to the virtual machine.

The previous virtual machine block (e.g. the entry block) will always push the operands for the next instruction onto the stack.

The handler for the instruction will read the operands and execute the original instruction.

If the instruction makes use of the rflags, then the flags are loaded/saved off of the context before and after the instruction.

If applicable, the result operand will then be pushed onto the stack if the instruction has one.

The next virtual machine block will read the result operand into the correct destination (if applicable, applying the changes to the stack-stored registers).

Handlers are added until an instruction that is not supported is found. The virtual machine will exit in this case.

When exiting the virtual machine,  all general purpose registers and the flags are popped.